requirements.txt
customtkinter>=5.2.0
openpyxl>=3.1.0
tkcalendar>=1.6.1
Pillow>=10.0.0
